poster-replacer/requirements.txt
rembg
opencv-python
numpy
Pillow
mediapipe
requests
